test: relocate frontend tests to top-level tests/ directory
