+++ tests/composables/SearchView/useModelDropdownPopup.test.ts
@@ -1,15 +1,15 @@
-import { popupManager, type SessionHistoryData } from '@services/PopupService';
+import { type ModelDropdownData, popupManager } from '@services/PopupService';
+import { mountComposable } from '@tests/utils/composables';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { mountComposable } from '../../../../tests/utils/composables';
-import { useSessionHistoryPopup } from './useSessionHistoryPopup';
+import { useModelDropdownPopup } from '@/views/SearchView/composables/useModelDropdownPopup';
 
 const { popupManagerMock, popupManagerState } = vi.hoisted(() => ({
     popupManagerState: {
         isOpen: true,
-        currentType: 'session-history-popup' as const,
-        currentPopupId: 'popup-session-history-popup:1',
-        currentWindowLabel: 'popup-session-history-popup',
+        currentType: 'model-dropdown-popup' as const,
+        currentPopupId: 'popup-model-dropdown-popup:1',
+        currentWindowLabel: 'popup-model-dropdown-popup',
         currentPopupSessionVersion: 1,
         isInitialized: true,
     },
@@ -21,12 +21,14 @@ const { popupManagerMock, popupManagerState } = vi.hoisted(() => ({
     },
 }));
 
-function createPopupData(overrides: Partial<SessionHistoryData> = {}): SessionHistoryData {
+function createPopupData(overrides: Partial<ModelDropdownData> = {}): ModelDropdownData {
     return {
-        sessions: [],
-        activeSessionId: null,
+        activeModelId: 'gpt-5',
+        activeProviderId: 1,
+        selectedModelId: 'gpt-5',
+        selectedProviderId: 1,
         searchQuery: '',
-        isLoading: false,
+        models: [],
         ...overrides,
     };
 }
@@ -38,18 +40,18 @@ vi.mock('@services/PopupService', () => ({
     },
 }));
 
-describe('useSessionHistoryPopup', () => {
+describe('useModelDropdownPopup', () => {
     beforeEach(() => {
         vi.clearAllMocks();
         Object.assign(popupManagerState, {
             isOpen: true,
-            currentType: 'session-history-popup',
-            currentPopupId: 'popup-session-history-popup:1',
-            currentWindowLabel: 'popup-session-history-popup',
+            currentType: 'model-dropdown-popup',
+            currentPopupId: 'popup-model-dropdown-popup:1',
+            currentWindowLabel: 'popup-model-dropdown-popup',
             currentPopupSessionVersion: 1,
             isInitialized: true,
         });
-        vi.mocked(popupManager.show).mockResolvedValue('popup-session-history-popup:1');
+        vi.mocked(popupManager.show).mockResolvedValue('popup-model-dropdown-popup:1');
         vi.mocked(popupManager.hide).mockResolvedValue(undefined);
         vi.mocked(popupManager.updateData).mockResolvedValue(undefined);
         vi.mocked(popupManager.listen).mockResolvedValue(() => undefined);
@@ -59,78 +61,72 @@ describe('useSessionHistoryPopup', () => {
         document.body.innerHTML = '';
     });
 
-    it('opens the popup and reports the active popup session only when popupManager confirms the same live session', async () => {
+    it('opens the popup and reports a live session when popupManager confirms the same session', async () => {
         const onPopupSessionStart = vi.fn();
-        const onPopupSessionEnd = vi.fn();
-        const anchor = document.createElement('button');
-
         const mounted = await mountComposable(() =>
-            useSessionHistoryPopup({
-                getAnchorElement: () => anchor,
-                getPopupData: () => createPopupData({ searchQuery: 'hello' }),
-                isSessionHistoryActive: () => true,
-                onSessionOpen: () => undefined,
-                onSessionSearchQueryChange: () => undefined,
+            useModelDropdownPopup({
+                getAnchorElement: () => document.createElement('button'),
+                getPopupData: () => createPopupData({ searchQuery: 'model' }),
+                isModelDropdownActive: () => true,
+                onModelSelect: () => undefined,
+                onModelSearchQueryChange: () => undefined,
                 onClose: () => undefined,
                 onPopupSessionStart,
-                onPopupSessionEnd,
             })
         );
 
         await mounted.result.open();
 
-        expect(popupManager.show).toHaveBeenCalledWith(
-            'session-history-popup',
-            anchor,
-            createPopupData({ searchQuery: 'hello' })
-        );
         expect(mounted.result.isOpen.value).toBe(true);
+        expect(mounted.result.isLiveSession()).toBe(true);
         expect(onPopupSessionStart).toHaveBeenCalledWith({
-            popupId: 'popup-session-history-popup:1',
-            windowLabel: 'popup-session-history-popup',
+            popupId: 'popup-model-dropdown-popup:1',
+            windowLabel: 'popup-model-dropdown-popup',
             popupSessionVersion: 1,
         });
-        expect(onPopupSessionEnd).not.toHaveBeenCalled();
 
         mounted.unmount();
     });
 
     it('rejects a popup session that is already stale by the time open resolves', async () => {
         Object.assign(popupManagerState, {
-            currentPopupId: 'popup-session-history-popup:2',
+            currentPopupId: 'popup-model-dropdown-popup:2',
             currentPopupSessionVersion: 2,
         });
         const onPopupSessionStart = vi.fn();
         const onPopupSessionEnd = vi.fn();
         const mounted = await mountComposable(() =>
-            useSessionHistoryPopup({
+            useModelDropdownPopup({
                 getAnchorElement: () => document.createElement('button'),
                 getPopupData: () => createPopupData({ searchQuery: 'stale-open' }),
-                isSessionHistoryActive: () => true,
-                onSessionOpen: () => undefined,
-                onSessionSearchQueryChange: () => undefined,
+                isModelDropdownActive: () => true,
+                onModelSelect: () => undefined,
+                onModelSearchQueryChange: () => undefined,
                 onClose: () => undefined,
                 onPopupSessionStart,
                 onPopupSessionEnd,
             })
         );
 
+        expect(mounted.result.isLiveSession()).toBe(false);
+
         await mounted.result.open();
 
         expect(mounted.result.isOpen.value).toBe(false);
+        expect(mounted.result.isLiveSession()).toBe(false);
         expect(onPopupSessionStart).not.toHaveBeenCalled();
         expect(onPopupSessionEnd).toHaveBeenCalledTimes(1);
 
         mounted.unmount();
     });
 
-    it('ignores stale popup close events and only notifies close for the active session', async () => {
+    it('ignores stale popup close events and preserves the live session until the active popup closes', async () => {
         let onCloseListener:
             | ((payload: {
                   popupId: string;
                   popupSessionVersion: number;
                   windowLabel: string;
-                  type: 'session-history-popup';
+                  type: 'model-dropdown-popup';
               }) => void)
             | undefined;
         vi.mocked(popupManager.listen).mockImplementation(async (handlers) => {
@@ -140,12 +136,12 @@ describe('useSessionHistoryPopup', () => {
         const onClose = vi.fn();
         const onPopupSessionEnd = vi.fn();
         const mounted = await mountComposable(() =>
-            useSessionHistoryPopup({
+            useModelDropdownPopup({
                 getAnchorElement: () => document.createElement('button'),
                 getPopupData: () => createPopupData(),
-                isSessionHistoryActive: () => true,
-                onSessionOpen: () => undefined,
-                onSessionSearchQueryChange: () => undefined,
+                isModelDropdownActive: () => true,
+                onModelSelect: () => undefined,
+                onModelSearchQueryChange: () => undefined,
                 onClose,
                 onPopupSessionEnd,
             })
@@ -154,20 +150,20 @@ describe('useSessionHistoryPopup', () => {
         await mounted.result.open();
 
         onCloseListener?.({
-            popupId: 'popup-session-history-popup:999',
-            popupSessionVersion: 999,
-            windowLabel: 'popup-session-history-popup',
-            type: 'session-history-popup',
+            popupId: 'popup-model-dropdown-popup:99',
+            popupSessionVersion: 99,
+            windowLabel: 'popup-model-dropdown-popup',
+            type: 'model-dropdown-popup',
         });
 
         expect(mounted.result.isOpen.value).toBe(true);
         expect(onClose).not.toHaveBeenCalled();
 
         onCloseListener?.({
-            popupId: 'popup-session-history-popup:1',
+            popupId: 'popup-model-dropdown-popup:1',
             popupSessionVersion: 1,
-            windowLabel: 'popup-session-history-popup',
-            type: 'session-history-popup',
+            windowLabel: 'popup-model-dropdown-popup',
+            type: 'model-dropdown-popup',
         });
 
         expect(mounted.result.isOpen.value).toBe(false);
@@ -177,13 +173,13 @@ describe('useSessionHistoryPopup', () => {
         mounted.unmount();
     });
 
-    it('closes the local popup session without notifying the page when session history is no longer active', async () => {
+    it('closes the local popup session without notifying the page when the model dropdown is no longer active', async () => {
         let onCloseListener:
             | ((payload: {
                   popupId: string;
                   popupSessionVersion: number;
                   windowLabel: string;
-                  type: 'session-history-popup';
+                  type: 'model-dropdown-popup';
               }) => void)
             | undefined;
         vi.mocked(popupManager.listen).mockImplementation(async (handlers) => {
@@ -193,12 +189,12 @@ describe('useSessionHistoryPopup', () => {
         const onClose = vi.fn();
         const onPopupSessionEnd = vi.fn();
         const mounted = await mountComposable(() =>
-            useSessionHistoryPopup({
+            useModelDropdownPopup({
                 getAnchorElement: () => document.createElement('button'),
                 getPopupData: () => createPopupData(),
-                isSessionHistoryActive: () => false,
-                onSessionOpen: () => undefined,
-                onSessionSearchQueryChange: () => undefined,
+                isModelDropdownActive: () => false,
+                onModelSelect: () => undefined,
+                onModelSearchQueryChange: () => undefined,
                 onClose,
                 onPopupSessionEnd,
             })
@@ -207,10 +203,10 @@ describe('useSessionHistoryPopup', () => {
         await mounted.result.open();
 
         onCloseListener?.({
-            popupId: 'popup-session-history-popup:1',
+            popupId: 'popup-model-dropdown-popup:1',
             popupSessionVersion: 1,
-            windowLabel: 'popup-session-history-popup',
-            type: 'session-history-popup',
+            windowLabel: 'popup-model-dropdown-popup',
+            type: 'model-dropdown-popup',
         });
 
         expect(mounted.result.isOpen.value).toBe(false);
@@ -220,25 +216,32 @@ describe('useSessionHistoryPopup', () => {
         mounted.unmount();
     });
 
-    it('hides the current popup session on unmount when the popup is still open', async () => {
+    it('closes the active popup session with the current identity', async () => {
         const onPopupSessionEnd = vi.fn();
         const mounted = await mountComposable(() =>
-            useSessionHistoryPopup({
+            useModelDropdownPopup({
                 getAnchorElement: () => document.createElement('button'),
                 getPopupData: () => createPopupData(),
-                isSessionHistoryActive: () => true,
-                onSessionOpen: () => undefined,
-                onSessionSearchQueryChange: () => undefined,
+                isModelDropdownActive: () => true,
+                onModelSelect: () => undefined,
+                onModelSearchQueryChange: () => undefined,
                 onClose: () => undefined,
                 onPopupSessionEnd,
             })
         );
 
         await mounted.result.open();
-        mounted.unmount();
+        await mounted.result.close();
 
-        expect(popupManager.hide).toHaveBeenCalledTimes(1);
-        expect(onPopupSessionEnd).toHaveBeenCalled();
+        expect(popupManager.hide).toHaveBeenCalledWith({
+            popupId: 'popup-model-dropdown-popup:1',
+            popupSessionVersion: 1,
+            windowLabel: 'popup-model-dropdown-popup',
+        });
+        expect(mounted.result.isOpen.value).toBe(false);
+        expect(onPopupSessionEnd).toHaveBeenCalledTimes(1);
+
+        mounted.unmount();
     });
 
     it('cleans up a late popup listener registration if the composable unmounts first', async () => {
@@ -252,12 +255,12 @@ describe('useSessionHistoryPopup', () => {
         );
 
         const mounted = await mountComposable(() =>
-            useSessionHistoryPopup({
+            useModelDropdownPopup({
                 getAnchorElement: () => document.createElement('button'),
                 getPopupData: () => createPopupData(),
-                isSessionHistoryActive: () => true,
-                onSessionOpen: () => undefined,
-                onSessionSearchQueryChange: () => undefined,
+                isModelDropdownActive: () => true,
+                onModelSelect: () => undefined,
+                onModelSearchQueryChange: () => undefined,
                 onClose: () => undefined,
             })
         );
